Split entity bridge module
